refactor(ui): split TLS certificates page into per-tab panels
The view had grown past 1500 lines. Move the certificates tab into
TlsCertificatesPanel and rename AcmeSettings to AcmeSettingsPanel, leaving
the view with the page chrome, the tab state and the shared Traefik instance
list it feeds to both panels.

The node filter stays bound to a query param, so the view owns it and passes
it down with .sync. Drop the selectedNodeLabel computed, which was never
referenced.

Also narrow the tooltip cursor rule: carbon already sets a pointer on the
tooltip trigger, only .bx--tag overrides it.
